requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.2.0
albumentations==1.3.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
bcrypt==4.0.1
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==39.0.0
cycler==0.11.0
decorator==4.4.2
efficientnet-pytorch==0.6.3
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
idna-ssl==1.1.0
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mean-average-precision==2021.4.26.0
multidict==5.2.0
munch==2.5.0
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.2
olefile==0.46
opencv-python-headless==4.7.0.68
packaging==21.3
pandas==1.1.5
paramiko==3.0.0
Pillow==8.4.0
pretrainedmodels==0.7.4
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.5.10
pytz==2022.7.1
PyWavelets==1.1.1
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
scp==0.14.5
segmentation-models-pytorch==0.2.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.4.12
tk==0.1.0
torch==1.10.2
torchaudio==0.10.2
torchmetrics==0.8.2
torchvision==0.11.3
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.0.3
yarl==1.7.2
zipfile36==0.1.3
zipp==3.6.0
